Fix overflow issue in PSNR computation
git-svn-id: svn://hevc.kw.bbc.co.uk/svn/jctvc-tmuc/branches/HM-1.0-dev@566 ce97f6c2-8f86-0410-b44d-af84b13ec34f

+++ source/Lib/TLibEncoder/TEncGOP.cpp
@@ -651,9 +651,9 @@ UInt64 TEncGOP::xFindDistortionFrame (TComPicYuv* pcPic0, TComPicYuv* pcPic1)
 Void TEncGOP::xCalculateAddPSNR( TComPic* pcPic, TComPicYuv* pcPicD, UInt uibits, Double dEncTime )
 {
   Int     x, y;
-  UInt    uiSSDY  = 0;
-  UInt    uiSSDU  = 0;
-  UInt    uiSSDV  = 0;
+  UInt64 uiSSDY  = 0;
+  UInt64 uiSSDU  = 0;
+  UInt64 uiSSDV  = 0;
   
   Double  dYPSNR  = 0.0;
   Double  dUPSNR  = 0.0;
